Refactor video
