+++ src/_sdl2/video/video.pyi
